refactor(example): update example with getStatus and getOutput calls in one single file
